Refactor directories to fix make
