+++ internal/buildflags/globals.go
@@ -1,0 +1,6 @@
+package buildflags
+
+var (
+	// Version is the git sha
+	Version string
+)
