+++ internal/controllers/interface/types.go
@@ -1,6 +1,9 @@
-package web
+package controllers
 
-import "plex_monitor/internal/database/models"
+import (
+	"plex_monitor/internal/buildflags"
+	"plex_monitor/internal/database/models"
+)
 
 // AppData is the data passed to the template.
 type AppData struct {
@@ -8,18 +11,22 @@ type AppData struct {
 	Name string
 	// User is the user that is logged in
 	User models.User
+	// Current revision of the app
+	Version string
 }
 
 // GetAppData returns the AppData struct.
 func GetAppData(user *models.User) AppData {
 	if user == nil {
 		return AppData{
-			Name: "Plex Monitor",
+			Name:    "Plex Monitor",
+			Version: buildflags.Version,
 		}
 	}
 
 	return AppData{
-		Name: "Plex Monitor",
-		User: *user,
+		Name:    "Plex Monitor",
+		User:    *user,
+		Version: buildflags.Version,
 	}
 }
